+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.1 under development
 
+- New #329: Add caption support for GridView (@Roc755)
 - Enh #328: Explicitly import constants in "use" section (@mspirkov)
 - Enh #331: Bump minimal `yiisoft/html` version to `3.13` and add support for `^4.0` (@vjik)